Addresses Seb's comments.  Utils reorganizing.
