+++ geoapps/inversion/electricals/application.py
@@ -58,8 +58,8 @@ from geoapps.base.plot import PlotSelection2D
 from geoapps.base.selection import ObjectDataSelection, TopographyOptions
 from geoapps.inversion.electricals import DirectCurrentParams, InducedPolarizationParams
 from geoapps.inversion.electricals.direct_current.constants import app_initializer
-from geoapps.utils.general import find_value
-from geoapps.utils.io import string_2_list
+from geoapps.utils.list import find_value
+from geoapps.utils.string import string_2_list
 
 
 def inversion_defaults():
