+++ geoapps/inversion/potential_fields/application.py
@@ -22,8 +22,8 @@ from geoapps.base.plot import PlotSelection2D
 from geoapps.base.selection import ObjectDataSelection, TopographyOptions
 from geoapps.inversion.potential_fields.magnetic_vector.constants import app_initializer
 from geoapps.utils import geophysical_systems, soft_import
-from geoapps.utils.general import find_value
-from geoapps.utils.io import string_2_list
+from geoapps.utils.list import find_value
+from geoapps.utils.string import string_2_list
 
 widgets = soft_import("ipywidgets")
 (
