+++ geoapps/iso_surfaces/application.py
@@ -15,7 +15,7 @@ from geoapps.utils import soft_import
 
 from geoapps.base.selection import ObjectDataSelection, TopographyOptions
 from geoapps.utils.formatters import string_name
-from geoapps.utils.general import input_string_2_float
+from geoapps.utils.plotting import input_string_2_float
 
 from .utils import iso_surface
